MAINT: move some helpers to test_utils
